smok tests are automagically carried out along with __gtg
